Swap all vectors to have doubles

+++ src/main/java/frc/team6718/vector/Vector2.java
@@ -1,9 +1,9 @@
 package frc.team6718.vector;
 
 public class Vector2 implements Vector<Vector2> {
-    public float x, y;
+    public double x, y;
 
-    public Vector2(float x, float y) {
+    public Vector2(double x, double y) {
         set(x,y);
     }
 
@@ -14,25 +14,25 @@ public class Vector2 implements Vector<Vector2> {
         return this;
     }
 
-    public Vector2 set(float x, float y) {
+    public Vector2 set(double x, double y) {
         this.x = x;
         this.y = y;
         return this;
     }
 
     @Override
-    public float magnitude() {
-        return (float) Math.sqrt(x * x  + y * y);
+    public double magnitude() {
+        return Math.sqrt(x * x  + y * y);
     }
 
     @Override
-    public float magnitude2() {
+    public double magnitude2() {
         return x * x * y * y;
     }
 
     @Override
     public Vector2 normalize() {
-        float length = magnitude();
+        double length = magnitude();
         x /= length;
         y /= length;
         return this;
@@ -45,7 +45,7 @@ public class Vector2 implements Vector<Vector2> {
         return this;
     }
 
-    public Vector2 add(float x, float y) {
+    public Vector2 add(double x, double y) {
         this.x = x;
         this.y = y;
         return this;
@@ -58,14 +58,14 @@ public class Vector2 implements Vector<Vector2> {
         return this;
     }
 
-    public Vector2 sub(float x, float y) {
+    public Vector2 sub(double x, double y) {
         this.x -= x;
         this.y -= y;
         return this;
     }
 
     @Override
-    public Vector2 mul(float v) {
+    public Vector2 mul(double v) {
         x *= v;
         y *= v;
         return this;
@@ -78,7 +78,7 @@ public class Vector2 implements Vector<Vector2> {
         return this;
     }
 
-    public Vector2 mul(float x, float y) {
+    public Vector2 mul(double x, double y) {
         this.x *= x;
         this.y *= y;
         return this;
